Fix: avoid refresh loop in diff views when stdout decode fails
When running diff/inline-diff on a binary file, both utf-8 and fallback
encoding would fail. When the fallback dialog is dismissed, the view regains
focus and is refreshed in the related event listener `on_activated` method;
which in turn would call the command again, and raise the same exception
when decode fails again.

If we add any other "refresh view" commands, we need to take care to check
for "git_savvy.decode_stdout_error" in the view settings if we depend on
succesful stdout content decoding.

Fix #579